Voucher implementation Tests passing

+++ database/migrations/2022_07_15_124501_create_order_product_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             
             $table->unsignedInteger('quantity');
             
-            $table->float('amount_in_pence');
+            $table->integer('amount_in_pence');
             $table->float('vat_rate');
         });
     }
